fix for 64 bit applied. test client fixed
git-svn-id: https://core.forge.funambol.org/svn/core/top-level/trunk/3x/client-api/native@19938 e8e8ed6c-164c-0410-afcf-9e9a7c7d8c10

+++ src/c++/common/client/DMTClientConfig.cpp
@@ -720,7 +720,7 @@ void DMTClientConfig::saveExtAccessConfig(ManagementNode& syncMLNode,
                                           ManagementNode& extNode) {
     char buf[512];
 
-    sprintf(buf, "%lu", accessConfig.getFirstTimeSyncMode());
+    sprintf(buf, "%u", (unsigned)accessConfig.getFirstTimeSyncMode());
     extNode.setPropertyValue(PROPERTY_FIRST_TIME_SYNC_MODE, buf);
 
     sprintf(buf, "%lu", accessConfig.getMaxMsgSize());
